image-captioning/requirements.txt
torch
transformers
Pillow
huggingface-hub
einops
timm